[Inventory] 재고 도메인 모델 및 Flyway 마이그레이션 (#51)
* [Inventory] 재고 도메인 모델 및 Flyway 마이그레이션

- 도메인 모델: Lot, InventoryLot, InventoryTransaction (참조는 Long xxxId, ADR-005)
- enum: LotStatus, QualityStatus, TransactionType, ReferenceType
- InventoryLot 수량 규칙(0 <= allocated <= onHand)과 증가/차감/할당/해제/출고 메서드
- V4 마이그레이션: lot, inventory_lot, inventory_transaction
  - lot.supplier_id, inventory_transaction.created_by는 참조 테이블 생성 전이라 FK 없이 인덱스만 생성
  - inventory_lot 수량 CHECK 제약 추가

* [Product] 안전 재고 수량 타입 BIGINT로 변경

- ERD 결정 사항(수량 컬럼은 모두 BIGINT)에 맞춰 product_sku.safety_stock_quantity를 DECIMAL(14,3)에서 BIGINT로 변경
- 적용된 V2 체크섬 보호를 위해 V5 마이그레이션으로 분리
- ProductSku 도메인, JPA 엔티티, 등록 요청/커맨드, 응답 DTO, 테스트의 타입을 BigDecimal에서 Long으로 변경

+++ src/main/java/com/kb/wms/product/adapter/in/web/dto/request/ProductSkuRegisterRequest.java
@@ -41,7 +41,7 @@ public record ProductSkuRegisterRequest(
         String unit,
 
         @PositiveOrZero(message = "안전 재고 수량은 0 이상이어야 합니다.")
-        BigDecimal safetyStockQuantity
+        Long safetyStockQuantity
 ) {
 
     public ProductSkuRegisterCommand toCommand() {
