+++ src/main/java/com/kb/wms/product/adapter/in/web/dto/response/ProductSkuListItemResponse.java
@@ -20,7 +20,7 @@ public record ProductSkuListItemResponse(
         BigDecimal weight,
         BigDecimal currentPurchasePrice,
         BigDecimal currentSupplyPrice,
-        BigDecimal safetyStockQuantity,
+        Long safetyStockQuantity,
         boolean isActive,
         List<SkuOptionValueResponse> optionValues
 ) {
